requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
python-dotenv==1.0.0
Flask-Mail==0.9.1